requirements.txt
# Voice Agent Dependencies
openai>=1.0.0
crewai>=0.70.0
elevenlabs>=1.0.0
mem0ai>=0.1.0
pyaudio>=0.2.11
python-dotenv>=1.0.0
pygame>=2.0.0

# Built-in modules (no installation needed):
# wave, tempfile, os

fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
twilio==8.10.0